+++ changelog.md
@@ -252,6 +252,8 @@ provided by the operating system.
 
 - `--newruntime` and `--refchecks` are deprecated.
 
+- Added `unsafeIsolate` and `extract` to `std/isolation`.
+
 ## Tool changes
 
 - The rst parser now supports markdown table syntax.